T09
In this task I have to Create a Python file called courier.py in this folder.
● You need to design a program for a courier company to calculate the cost
of sending a parcel.
● Ask the user to enter the price of the package they would like to purchase.
● Ask the user to enter the total distance of the delivery in kms.
● Now add on the delivery costs to get the final cost of the product:
o There are four categories to factor in when determining a parcel’s
final cost, each with two options based on their delivery
preferences. (Use an if-else statement based on the choice they
make.)
▪ Air R0.36 per km or freight R0.25 per km
▪ Full insurance R50.00 or limited insurance R25.00
▪ Gift R15.00 or no gift R0.00
▪ Priority R100.00 or standard delivery R20.00
● Work out the total cost of the package based on the selection in each
category.